The Local Peace Agreements analysis from Shin
Analysis: To have a thorough visualization on each countries that were mentioned in the local peace agreement in terms of their location determined by longitude and latitude, there are two graphs individually being visualized based on the longitude and the Latitude variables. As we observe from these graphs, the majority of the countries are located around the equator.