The concept of graphics visualization
These are the design and the overall process of showing each graphs I created for our website. In this files it has included the individual reports from Xia, QU and me. I was doing some researched on trying to portraits the culture of Syria which is our main focus of the website by implementing some of their iconic-local elements. Although I am struggling with finding the balance in designing the proportion of each graphs, the color themes and other issues that has been raised during my design process. I have opened an issue section and waiting for feedback from teammates.