ReadMe
There are two versions of the visualization for our graphics. One in which I arised three issues for looking for feedback from teammates. The 2nd version is the modifications I re-designed completely based on the feedbacks from the team. As well as following the color schemes provided by the teams.