Final report
Our project is building an interactive website centered on data extraction from the Local Peace Agreement. Moreover, the target audiences we are aiming at are socialists, researchers, potential mediators and peace-loving people. The information about our analysis content can be exclaimed into two major sections, the overview and the individual reports, in which the later one can be found on the website that we created. While in this report, we would like to focus on some of the researches we investigated in the overview section.