local peace agreement
This database lists all agreements included on the main PA-X database that deal in some way with local issues, in cases of local negotiation practices that are on the margins of wider conflict, or in cases of local/communal violent conflict. The detailed database covers 25 countries in the Asia-Pacific region, Middle East and North Africa, Europe and Eurasia, Africa（excl MENA） and Cross-regional five regions, with a total of about 286 agreements between 1990 and 2020.The dataset are settled into three parts, basic information, local agreement properties, and local agreement issues.