Analysis of the influence of culture and economy on mediators
We mainly analyzed the influence of culture and economy on the types of mediators. 
In terms of culture, because one of the process types is Informal but persistent practice (PT_IPP), it represents a clear culture or practice of local peace agreement-making, so we use PT_IPP to verify the influence of culture on the mediator type.
In terms of economy, we use GDP as a symbol of economic strength. We mainly inquired about countries that signed peace agreements between 2016 and 2018 and used this to analyze the impact of the economy on the types of mediators.